# api-spec.yaml
swagger: '2.0'
info:
  title: Blockstore API
  description: REST API for Open edX Blockstore
  license:
    name: AGPL
  version: v1
host: localhost:18250
schemes:
  - http
basePath: /
consumes:
  - application/json
produces:
  - application/json
securityDefinitions: {}
paths:
...
  /tagstore/api/v1/entities/{entity_type}/{_id}:
    get:
      operationId: retrieve_entity
      description: "Get a single entity. Never raises a 404, because Tagstore doesn't\
        \ know\nwhich entities exist or not. If you want to know whether or not the\n\
        entity is persisted in Tagstore's database, check the resulting\n\"persisted\"\
        \ boolean field."
      parameters: []
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/EntityDetail'
      tags:
        - tagstore
    parameters:
      - name: _id
        in: path
        description: A unique integer value identifying this entity.
        required: true
        type: integer
      - name: entity_type
        in: path
        required: true
        type: string
...
  /tagstore/api/v1/taxonomies:
    get:
      operationId: get_taxonomies
      description: Get a list of all taxonomies.
      parameters:
        - name: page
          in: query
          description: A page number within the paginated result set.
          required: false
          type: integer
        - name: page_size
          in: query
          description: Number of results to return per page.
          required: false
          type: integer
      responses:
        '200':
          description: ''
          schema:
            required:
              - count
              - results
            type: object
            properties:
              count:
                type: integer
              next:
                type: string
                format: uri
                x-nullable: true
              previous:
                type: string
                format: uri
                x-nullable: true
              results:
                type: array
                items:
                  $ref: '#/definitions/Taxonomy'
      tags:
        - tagstore
    post:
      operationId: create_taxonomy
      description: Create a new Taxonomy
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/Taxonomy'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/Taxonomy'
      tags:
        - tagstore
    parameters: []
  /tagstore/api/v1/taxonomies/{id}:
    get:
      operationId: get_taxonomy
      description: Get a specific taxonomy.
      parameters: []
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/Taxonomy'
      tags:
        - tagstore
    delete:
      operationId: delete_taxonomy
      description: Delete a specific taxonomy and all of its tags.
      parameters: []
      responses:
        '204':
          description: ''
      tags:
        - tagstore
    parameters:
      - name: id
        in: path
        description: A unique integer value identifying this taxonomy.
        required: true
        type: integer
  /tagstore/api/v1/taxonomies/{id}/tags:
    get:
      operationId: get_taxonomy_tags
      summary: List the tags in this taxonomy.
      description: It guarantees that parent tags will be returned before their children.
      parameters:
        - name: page
          in: query
          description: A page number within the paginated result set.
          required: false
          type: integer
        - name: page_size
          in: query
          description: Number of results to return per page.
          required: false
          type: integer
      responses:
        '200':
          description: ''
          schema:
            required:
              - count
              - results
            type: object
            properties:
              count:
                type: integer
              next:
                type: string
                format: uri
                x-nullable: true
              previous:
                type: string
                format: uri
                x-nullable: true
              results:
                type: array
                items:
                  $ref: '#/definitions/TagWithHierarchy'
      tags:
        - tagstore
    post:
      operationId: add_taxonomy_tag
      description: Add a tag to the taxonomy, if it doesn't already exist
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/Tag'
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/TagWithHierarchy'
      tags:
        - tagstore
    parameters:
      - name: id
        in: path
        description: A unique integer value identifying this taxonomy.
        required: true
        type: integer
  /tagstore/api/v1/taxonomies/{id}/tags/{tag_name}:
    get:
      operationId: get_taxonomy_tag
      description: Get a specific tag in the taxonomy
      parameters: []
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/TagWithHierarchy'
      tags:
        - tagstore
    delete:
      operationId: delete_taxonomy_tag
      description: Delete a tag
      parameters: []
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/EmptyObject'
      tags:
        - tagstore
    parameters:
      - name: id
        in: path
        description: A unique integer value identifying this taxonomy.
        required: true
        type: integer
      - name: tag_name
        in: path
        required: true
        type: string
definitions:
...
  Tag:
    required:
      - name
    type: object
    properties:
      taxonomy_id:
        title: Taxonomy id
        type: string
        readOnly: true
      name:
        title: Name
        description: Display name of this tag; also its identifier within the taxonomy.
        type: string
        minLength: 1
      path:
        title: Path
        description: "\n        'path' is provided for use in\n        implementing\
          \ searching by tag. It provides the following guarantees:\n        (1) 'path'\
          \ is globally unique.\n        (2) any child tag's path string starts with\
          \ its parent's tag's path.\n    "
        type: string
        readOnly: true
        minLength: 1
  EntityDetail:
    required:
      - entity_type
      - external_id
    type: object
    properties:
      entity_type:
        title: Entity type
        type: string
        minLength: 1
      external_id:
        title: External id
        type: string
        minLength: 1
      persisted:
        title: Persisted
        type: boolean
        readOnly: true
      tags:
        description: "\n        The list of tags that this entity has applied.\n \
          \       "
        type: array
        items:
          $ref: '#/definitions/Tag'
        readOnly: true
  EmptyObject:
    type: object
    properties: {}
  Entity:
    title: Owner
    required:
      - entity_type
      - external_id
    type: object
    properties:
      entity_type:
        title: Entity type
        type: string
        minLength: 1
      external_id:
        title: External id
        type: string
        minLength: 1
  Taxonomy:
    required:
      - name
    type: object
    properties:
      id:
        title: Id
        type: integer
        readOnly: true
      name:
        title: Name
        type: string
        maxLength: 180
        minLength: 1
      owner:
        $ref: '#/definitions/Entity'
  TagWithHierarchy:
    required:
      - name
    type: object
    properties:
      taxonomy_id:
        title: Taxonomy id
        type: string
        readOnly: true
      name:
        title: Name
        description: Display name of this tag; also its identifier within the taxonomy.
        type: string
        minLength: 1
      path:
        title: Path
        description: "\n        'path' is provided for use in\n        implementing\
          \ searching by tag. It provides the following guarantees:\n        (1) 'path'\
          \ is globally unique.\n        (2) any child tag's path string starts with\
          \ its parent's tag's path.\n    "
        type: string
        readOnly: true
        minLength: 1
      parent:
        title: Parent
        type: string
        minLength: 1
        x-nullable: true